+++ crates/sampo-core/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "sampo-core"
+version = "0.1.0"
+license = "MIT"
+authors = ["Goulven Clech <goulven.clech@protonmail.com>"]
+edition = "2024"
+description = "Core types and utilities for Sampo - Internal library, not published"
+homepage = "https://github.com/bruits/sampo"
+repository = "https://github.com/bruits/sampo"
+readme = "README.md"
+keywords = ["changeset", "versioning", "publishing", "semver", "monorepo"]
+categories = ["development-tools"]
+publish = false
+
+[dependencies]
+toml = "0.8"
+glob = "0.3"
+tempfile = { version = "3.13", optional = true }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+thiserror = "1.0"
+tokio = { version = "1.0", features = ["rt", "macros"] }
+
+[dev-dependencies]
+tempfile = "3.13"
+
+[features]
+default = []
+testing = ["tempfile"]
