+++ crates/sampo-core/README.md
@@ -1,0 +1,13 @@
+# Sampo Core
+
+Provides core logic, common types, and internal utilities shared across all Sampo crates.
+
+This library is not published to crates.io and is only used internally within the Sampo workspace.
+
+## Features
+
+- **Types**: Common data structures (`CrateInfo`, `Workspace`, `Config`, etc.)
+- **Workspace**: Logic for discovering and parsing Cargo workspaces
+- **Changesets**: Utilities for reading and manipulating changeset files
+- **Configuration**: Shared configuration handling
+- **Git/GitHub**: Utilities for Git and GitHub integration
