+++ crates/sampo/src/main.rs
@@ -1,13 +1,9 @@
 mod add;
-mod changeset;
 mod cli;
-mod config;
-mod enrichment;
 mod init;
 mod names;
 mod publish;
 mod release;
-mod workspace;
 
 use clap::Parser;
 use cli::{Cli, Commands};
